% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{bin_mode}
\alias{bin_mode}
\title{Calculate the mode of binned variable}
\usage{
bin_mode(var, n_bins, data)
}
\arguments{
\item{var}{The variable}

\item{n_bins}{The number of bins with which to bin up the variable}

\item{data}{The data set}
}
\value{
The mid point of the bin that occurs with the highest frequency.
}
\description{
Calculate the mode of binned variable
}
\examples{
bin_mode(log_rt, n_bins = 5, data = behmercrump)
}
