% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{binomial_likelihood_plot}
\alias{binomial_likelihood_plot}
\title{Plot the likelihood function of a binomial model}
\usage{
binomial_likelihood_plot(m, n, lim = c(0, 1))
}
\arguments{
\item{m}{Number of successes}

\item{n}{Number of trials}
}
\value{
A ggplot object
}
\description{
Plot the likelihood function of a binomial model
}
\examples{
binomial_likelihood(250, 135)
}
