% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{binomial_posterior_plot}
\alias{binomial_posterior_plot}
\title{Plot the posterior distribution of a binomial model with Beta prior}
\usage{
binomial_posterior_plot(
  n,
  m,
  alpha = 1,
  beta = 1,
  show_hpd = TRUE,
  level = 0.95,
  xlim = c(0, 1)
)
}
\arguments{
\item{n}{Number of trials}

\item{m}{Number of successes}

\item{alpha}{First shape parameters of the Beta prior}

\item{beta}{Second shape parameter of the Beta prior}

\item{show_hpd}{Show the HPD interval}

\item{level}{The amount of mass in HPD interval}

\item{xlim}{The limits on x-axis}
}
\value{

}
\description{
Plot the posterior distribution of a binomial model with Beta prior
}
\examples{
binomial_posterior_plot(250, 139, 5, 5)
binomial_posterior_plot(100, 60, 2, 2, level = 0.99)
}
