% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{binomial_posterior_summary}
\alias{binomial_posterior_summary}
\title{Summary stats of posterior distribution of binomial model with Beta prior}
\usage{
binomial_posterior_summary(sample_size, observed, alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{First shape parameter of the Beta prior}

\item{beta}{Second shape parameter of the Beta prior}

\item{n}{Number of trials}

\item{m}{Number of successes}
}
\value{
A list with summary statistics of the posterior distribution
}
\description{
Summary stats of posterior distribution of binomial model with Beta prior
}
\examples{
binomial_posterior_summary(3, 5)
}
