% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{binomial_pvalue}
\alias{binomial_pvalue}
\title{Calculate p-value for a hypothesis in binomial problem}
\usage{
binomial_pvalue(sample_size, observed, hypothesis)
}
\arguments{
\item{sample_size}{The sample size.}

\item{observed}{The observed number of successes in the sample.}

\item{hypothesis}{The hypothesized probability parameter.}
}
\value{

}
\description{
Assuming that we have observed a certain number of successes from a binomial
distribution, obtain the p-value for a hypothesis about the binomial
distribution probability parameter, i.e. the probability of a success on each
trial.
}
\examples{
# p-value for hypothesis that prob=0.4, given 3 successes in 10 trials
binomial_pvalue(10, 3, 0.4)
}
