% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{get_beta_hpd}
\alias{get_beta_hpd}
\title{The high posterior density interval of a Beta distribution.}
\usage{
get_beta_hpd(alpha, beta, level = 0.95)
}
\arguments{
\item{alpha}{First shape parameter of the Beta distribution.}

\item{beta}{Second shape parameter of the Beta distribution.}
}
\value{
A list with lower and upper bound of the HPD interval, and minimum density of interval.
}
\description{
The high posterior density interval of a Beta distribution.
}
\examples{
get_beta_hpd(10, 5)
}
