% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{get_mode}
\alias{get_mode}
\title{Find the mode of a discrete (categorical or integer-like) variable}
\usage{
get_mode(data, var)
}
\arguments{
\item{data}{A data frame.}

\item{var}{A column of \code{data}.}
}
\value{
A tibble with two columns:
\describe{
\item{\code{var}}{the modal value(s) of the \code{var}}
\item{\code{count}}{how many times each mode occurs}
}
}
\description{
\code{get_mode()} returns the most frequent value(s) of a variable in a data
frame.  If several values are tied for highest frequency, they are all
returned.  Missing values (\code{NA}) are removed before counting.
}
\examples{
## Mode of the three-level income factor
get_mode(whr2025, income)

}
