% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{get_residual_df}
\alias{get_residual_df}
\title{Residual degrees of freedom of a fitted linear model}
\usage{
get_residual_df(model)
}
\arguments{
\item{model}{An object of class \code{"lm"}.}
}
\value{
A single integer giving the residual degrees of freedom.
}
\description{
Returns the residual degrees of freedom, \eqn{n - p - 1}, stored in an
\code{\link[stats]{lm}} object.
This quantity equals the sample size minus the number of fitted
coefficients (intercept + slopes) and represents the amount of independent
information left for estimating the error variance.
}
\examples{
fit <- lm(mpg ~ wt + hp, data = mtcars)
get_residual_df(fit)

}
