% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{get_rsq}
\alias{get_rsq}
\title{Extract R-squared from a fitted linear model}
\usage{
get_rsq(model)
}
\arguments{
\item{model}{An object of class \code{"lm"}.}
}
\value{
A numeric scalar giving the in-sample \eqn{R^{2}}.
}
\description{
Retrieves the coefficient of determination (\eqn{R^{2}}) from an object
produced by \code{\link[stats]{lm}}.
}
\details{
Internally the function calls \code{\link[stats]{summary.lm}} and returns
the \code{r.squared} component.
A descriptive error is issued if the supplied object is not of class
\code{"lm"}.
}
\examples{
mod <- lm(mpg ~ wt + hp, data = mtcars)
get_rsq(mod) # 0.8268

}
