% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{get_rss}
\alias{get_rss}
\title{Residual sum of squares (RSS) for an \code{lm} object}
\usage{
get_rss(model)
}
\arguments{
\item{model}{An object of class \code{"lm"}.}
}
\value{
A single numeric value giving the residual sum of squares for the
fitted model.
}
\description{
Computes the residual sum of squares, \eqn{\sum_{i}\hat\varepsilon_i^{2}},
from a fitted linear model.
The quantity is sometimes called the error sum of squares.
}
\examples{
fit <- lm(mpg ~ wt + hp, data = mtcars)
get_rss(fit)

}
