% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{gg_epsilon}
\alias{gg_epsilon}
\title{The Greenhouse & Geisser sphericity correction factor}
\usage{
gg_epsilon(Sigma)
}
\arguments{
\item{Sigma}{A sample covariance matrix.}
}
\value{
The epsilon-hat sphericity correction factor.
}
\description{
Calculates the Greenhouse & Geisser epsilon-hat factor for correction of sphericity.
This appeared in Geisser, S., & Greenhouse, S. W. (1958), and Greenhouse & Geisser (1959).
}
