% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{hf_epsilon}
\alias{hf_epsilon}
\title{The Huynh-Feldt sphericity correction factor}
\usage{
hf_epsilon(Sigma, n)
}
\arguments{
\item{Sigma}{}

\item{n}{}
}
\description{
Calculates the Huynh-Feldt epsilon-tilde factor for correcting sphericity.
This appeared in Huynh & Feldt (1976).
}
