% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdsr-package.R
\docType{package}
\name{isdsr-package}
\alias{isdsr}
\alias{isdsr-package}
\title{isdsr: Tools and Datasets to Accompany ISDSR Textbook}
\description{
Tools and datasets that are used in the textbook "Introduction to Statistics and Data Science Using R" by Mark Andrews & Lucy V. Justice (forthcoming), SAGE.
}
\author{
\strong{Maintainer}: Mark Andrews \email{mark.andrews@ntu.ac.uk}

}
\keyword{internal}
