% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mathplacement}
\alias{mathplacement}
\title{Scores from a Mathematics Placement Exam}
\format{
A data frame with 2696 rows and 1 variable
\describe{
\item{\code{score}}{(numeric) Score on exam}
}
}
\source{
This data set is a subset of the \code{MathPlacement} data set in the \href{https://cran.r-project.org/package=Stat2Data}{\code{Stats2Data} R package}.
Specifically, the \code{score} variable is the the \code{PlcmtScore} variable from \code{MathPlacement}.
}
\description{
Scores on a mathematics placement exam in an unnaned liberal arts college, presumably in the US.
}
