% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalinference.R
\name{normal_inference_pvalue}
\alias{normal_inference_pvalue}
\title{Calculate the p-value for a hypothesis test about mean of normal distribution}
\usage{
normal_inference_pvalue(y, mu, sigma_known = FALSE, na.rm = T)
}
\arguments{
\item{y}{(numeric) A set of values assumed to be drawn independently from a normal distribution.}

\item{mu}{(numeric) The hypothesized mean of the normal distribution.}

\item{sigma_known}{(logical) Is sigma, the standard deviation of the normal distribution, known?}

\item{na.rm}{(logical) When calculating the sufficient statistics, do we remove NA's}
}
\value{
(numeric) A p-value for the hypothesis
}
\description{
Given a set of values assumed to sampled independently from a normal distribution,
calculate the p-value for any hypothetical value of the distribution's mean.
}
\details{
The standard deviation of the normal distribution can assumed to be known to have the value of the sample standard deviation.
If so, the p-value is calculated assuming a normal sampling distribution.

If the standard deviation is not assumed to be known, the p-value is calculated using the t-distribution.
}
\examples{
y <- rnorm(10, mean = 1)
normal_inference_pvalue(y, mu = 1)
}
