% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalinference.R
\name{normal_interval}
\alias{normal_interval}
\title{Prediction interval of a normal distribution}
\usage{
normal_interval(mean = 0, sd = 1, probability = 0.95)
}
\arguments{
\item{mean}{(numeric) The mean of the normal distribution}

\item{sd}{(numeric) The standard deviation of the normal distribution}

\item{probability}{(numeric) The probability (area under curve) within the interval}
}
\value{
A numeric vector giving the lower and upper bounds of the interval containing the specified probability.
}
\description{
This function calculates the (central) interval of a normal distribution that
contains a specified probability. This interval will always extend from a
certain distance below the mean to the same distance above the mean.
}
\examples{
normal_interval()
normal_interval(mean = 100, sd = 15, probability = 0.99)
}
