% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalinference.R
\name{qqnormplot}
\alias{qqnormplot}
\title{Plot a normal QQ-plot}
\usage{
qqnormplot(var, data)
}
\arguments{
\item{var}{The variable to be compared to the}

\item{data}{The data frame contain the variable}
}
\value{
A ggplot object
}
\description{
Given a set of values, the quantiles (percentiles) of each value is
calculated. The values in a standard normal corresponding to each quantile is
calculated. The z-scores of the values of the variable are then plotted
against their corresponding z-scores in the standard normal.
}
\examples{
qqnormplot(score, mathplacement)
}
