% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{re_scale}
\alias{re_scale}
\title{Z-score selected numeric columns}
\usage{
re_scale(data, ...)
}
\arguments{
\item{data}{A data frame / tibble.}

\item{...}{Tidy-select expression(s) that pick the columns to rescale.
If omitted, \emph{all} numeric columns are standardised.}
}
\value{
A data frame of the same class as \code{data}.
}
\description{
Z-score selected numeric columns
}
\examples{
# Format all numeric columns to 3 decimal places
mtcars_df <- tibble::as_tibble(mtcars)
re_scale(mtcars_df)

# Just selected columns
re_scale(mtcars_df, mpg:qsec)

}
