% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgsur-package.R
\docType{package}
\name{sgsur-package}
\alias{sgsur}
\alias{sgsur-package}
\title{sgsur: Tools and Datasets to Accompany SGSUR Textbook}
\description{
Tools and datasets that are used in the textbook "A Student's Guide to Statistics Using R" by Mark Andrews & Lucy V. Justice (2026), SAGE.
}
\author{
\strong{Maintainer}: Mark Andrews \email{mark.andrews@ntu.ac.uk}

}
\keyword{internal}
