% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{simple_main_effects}
\alias{simple_main_effects}
\title{Simple main effects Anova}
\usage{
simple_main_effects(formula, by, data, table_only = TRUE)
}
\arguments{
\item{formula}{The \code{aov_car} formula}

\item{by}{The independent variable for each of whose levels a separate Anova
is performed.}

\item{data}{The data set.}

\item{table_only}{If TRUE (default), only a data-frame of F-tests is
returned. If FALSE, a list with each \code{aov_car} model is returned.}
}
\value{
A list of \code{afex::afex_aov} objects
}
\description{
For each level of one independent variable, perform an Anova with another
independent variable.
}
\examples{
simple_main_effects(log_rt ~ Error(subject/condition), by = keyboard, data = behmercrump)
}
