% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{sprintf_fstat}
\alias{sprintf_fstat}
\title{Format an F-test for reporting}
\usage{
sprintf_fstat(model, digits = 2)
}
\arguments{
\item{model}{An object that inherits from class \code{lm} or \code{aov}.}

\item{digits}{A single integer giving the number of decimal places
to show for the F statistic.}
}
\value{
A length-one character vector like
\code{"F(4, 72) = 3.52, p = .011"}.
}
\description{
Produce an APA-style string containing the F statistic,
its numerator and denominator degrees of freedom,
and the corresponding p-value.
}
\examples{
mod <- lm(mpg ~ factor(cyl), data = mtcars)
sprintf_fstat(mod)
}
