% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalinference.R
\name{t_interval}
\alias{t_interval}
\title{Prediction interval of a t-distribution}
\usage{
t_interval(nu = 1, probability = 0.95)
}
\arguments{
\item{nu}{(numeric) The degrees of freedom parameter of the t-distribution.
This is usually, but not necessarily, an integer.}

\item{probability}{(numeric) The probability (area under curve) within the
interval}
}
\value{
A numeric vector giving the lower and upper bounds of the interval
containing the specified probability.
}
\description{
Prediction interval of a t-distribution
}
\examples{
t_interval()
t_interval(nu = 10, probability = 0.99)
}
