% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{urn_sampler}
\alias{urn_sampler}
\title{Sample red and black marbles from an urn of marbles #'}
\usage{
urn_sampler(
  red = 50,
  black = 50,
  sample_size = 25,
  replace = TRUE,
  repetitions = 1
)
}
\arguments{
\item{red}{The number of red marbles in the urn.}

\item{black}{The number of black marbles in the urn.}

\item{sample_size}{The number of marbles to draw from the urn.}

\item{replace}{Is the sampling with or without replacement}

\item{repetitions}{The number of repetitions of this sampling to be done.}
}
\value{
A data frame with two columns: \code{n_red} and \code{n_black}, which give the
number of red and black marbles, respectively, in the sample of size
\code{sample_size}. Each row of the data frame gives the results of the
repetition of this experiment.
}
\description{
Sample red and black marbles from an urn of marbles #'
}
\examples{
urn_sampler(repetitions = 5)
}
