% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{urn_sampling_distribution}
\alias{urn_sampling_distribution}
\title{The exact sampling distribution for an urn problem}
\usage{
urn_sampling_distribution(
  red = 50,
  black = 50,
  sample_size = 25,
  red_proportion = NULL,
  observed_red = NULL,
  replace = TRUE
)
}
\arguments{
\item{red}{The number of red marbles in the urn.}

\item{black}{The number of black marbles in the urn.}

\item{sample_size}{The number of marbles to draw from the urn.}

\item{red_proportion}{As an alternative to specifying the number of
red and black marbles in the jar, the proportion of red marbles
can be specified. If \code{red_proportion} is specified, then
\code{red} and \code{black} will be ignored.}

\item{observed_red}{The number of red marbles in a sample. This
is optional. TODO more documentation here please}

\item{replace}{Is the sampling with or without replacement?}
}
\value{

}
\description{
The exact sampling distribution for an urn problem
}
\examples{
urn_sampling_distribution(sample_size = 10)
}
