% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urns.R
\name{binomial_posterior_hpd}
\alias{binomial_posterior_hpd}
\title{The high posterior density interval in a binomial problem with Beta prior.}
\usage{
binomial_posterior_hpd(
  sample_size,
  observed,
  alpha = 1,
  beta = 1,
  level = 0.95
)
}
\arguments{
\item{alpha}{First shape parameter of the Beta prior distribution.}

\item{beta}{Second shape parameter of the Beta prior distribution.}

\item{m}{Number of successes}

\item{n}{Number of trials}
}
\value{
A vector with lower and upper bound of the HPD interval.
}
\description{
The high posterior density interval in a binomial problem with Beta prior.
}
\examples{
binomial_posterior_hpd(172, 1300)
}
