% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalinference.R
\name{normal_inference_confint}
\alias{normal_inference_confint}
\title{Confidence interval for the mean of normal distribution}
\usage{
normal_inference_confint(y, level = 0.95, sigma_known = FALSE, na.rm = T)
}
\arguments{
\item{y}{(numeric) A set of values assumed to be drawn independently from a normal distribution.}

\item{level}{(numeric) The level of the confidence interval.}

\item{sigma_known}{(logical) Is sigma, the standard deviation of the normal distribution, known?}

\item{na.rm}{(logical) When calculating the sufficient statistics, do we remove NA's}
}
\value{
(numeric) The lower and upper bounds of the confidence interval
}
\description{
Given a set of values assumed to sampled independently from a normal
distribution, calculate the confidence interval of the distribution's mean.
}
\details{
The standard deviation of the normal distribution can assumed to be known to
have the value of the sample standard deviation. If so, the confidence
interval is calculated assuming a normal sampling distribution.

If the standard deviation is not assumed to be known, the confidence interval
is calculated using the t-distribution.
}
\examples{
y <- rnorm(10, mean = 5)
normal_inference_confint(y)
}
