% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{skim}
\alias{skewness}
\alias{kurtosis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{moments}{\code{\link[moments]{kurtosis}}, \code{\link[moments]{skewness}}}

  \item{skimr}{\code{\link[skimr]{skim}}}
}}

