% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{skim}
\alias{skewness}
\alias{hedges_g}
\alias{cohens_d}
\alias{leveneTest}
\alias{tidy}
\alias{augment}
\alias{glance}
\alias{emmeans}
\alias{aov_car}
\alias{nice}
\alias{lmer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{afex}{\code{\link[afex]{aov_car}}, \code{\link[afex]{nice}}}

  \item{broom}{\code{\link[broom:reexports]{augment}}, \code{\link[broom:reexports]{glance}}, \code{\link[broom:reexports]{tidy}}}

  \item{car}{\code{\link[car]{leveneTest}}}

  \item{effectsize}{\code{\link[effectsize]{cohens_d}}, \code{\link[effectsize:cohens_d]{hedges_g}}}

  \item{emmeans}{\code{\link[emmeans]{emmeans}}}

  \item{lme4}{\code{\link[lme4]{lmer}}}

  \item{moments}{\code{\link[moments]{skewness}}}

  \item{skimr}{\code{\link[skimr]{skim}}}
}}

