% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{twoway_interaction_plot}
\alias{twoway_interaction_plot}
\title{Interaction plot using expected marginal means}
\usage{
twoway_interaction_plot(model, iv1, iv2, ylab, text_size = 8)
}
\arguments{
\item{model}{A model such as afex::aov_car of a two-way interaction.}

\item{iv1}{One of the two independent variables (variable name must be quoted)}

\item{iv2}{The other of the two independent variables (variable name must be quoted)}

\item{ylab}{The y-axis label}

\item{text_size}{The size of the font in the legend}
}
\value{
A \code{patchwork} object
}
\description{
Interaction plot using expected marginal means
}
\examples{
M <- afex::aov_car(nc_erp ~ category + Error(id / change), data = catknowledge)
twoway_interaction_plot(M, "category", "change", ylab = "Nc ERP amplitude")
}
